requirements.txt
bs4
lxml
selenium
requests
zxing
Pillow-PIL
pillow
qrcode
SQLAlchemy
PyMysqlDB
CookieJar